Refactor checking
